+++ apps/uniapp/src/pages/404.vue
@@ -1,0 +1,3 @@
+<template>
+  <view> 404 </view>
+</template>