+++ platforms/mobile/src/mui/components/x-context-menu/index.ts
@@ -1,3 +1,0 @@
-import XContextMenu from './x-context-menu.vue';
-export * from './types';
-export { XContextMenu };