+++ platforms/mobile/src/mui/components/x-image-viewer/index.ts
@@ -1,3 +1,0 @@
-import XImageViewer from './x-image-viewer.vue';
-export * from './types';
-export { XImageViewer };