+++ platforms/pro-uni/src/main.ts
@@ -9,16 +9,12 @@ import {
   parseFunction,
   type Provider
 } from '@vtj/renderer';
-import type { BlockSchema } from '@vtj/core';
-import {
-  setupUniApp,
-  createUniAppComponent,
-  createPreviewUniRoutes
-} from '@vtj/uni';
+import { setupUniApp, createUniAppComponent, createUniRoutes } from '@vtj/uni';
 import { notify, loading } from './shared';
 
 const adapter = createAdapter({ loading, notify });
 const service = new LocalService(createServiceRequest(notify));
+
 const { provider, onReady } = createProvider({
   nodeEnv: process.env.NODE_ENV as NodeEnv,
   mode: ContextMode.Runtime,
@@ -31,50 +27,15 @@ const { provider, onReady } = createProvider({
   }
 });
 
-const dsl: BlockSchema = {
-  name: 'UniPageDemo',
-  nodes: [
-    {
-      name: 'View',
-      children: [
-        {
-          name: 'Button',
-          children: 'Button'
-        }
-      ]
-    }
-  ],
-  lifeCycles: {
-    onLoad: {
-      type: 'JSFunction',
-      value: `
-      (opt)=>{
-      console.log('onLoad app',opt, uni)
-      }
-        `
-    },
-    onShow: {
-      type: 'JSFunction',
-      value: `
-      ()=>{
-      console.log('onShow app')
-      }
-        `
-    }
-  }
-};
-
-const init = (provider: Provider) => {
+const init = async (provider: Provider) => {
   const { Vue, UniH5 } = window as any;
   const project = provider.project;
   if (!project) return;
-  // const { renderer } = createRenderer({ dsl, components: {} });
 
-  const App = createUniAppComponent(project.uniConfig || {}, (script: any) =>
+  const App = createUniAppComponent(project.uniConfig || {}, (script) =>
     parseFunction(script, window, false, true)
   );
-
-  const routes = createPreviewUniRoutes(Vue, createRenderer);
+  const routes = await createUniRoutes(provider, createRenderer);
 
   const app = setupUniApp({
     Vue,
@@ -82,6 +43,7 @@ const init = (provider: Provider) => {
     UniH5,
     routes
   });
+
   app.use(provider);
   app.mount(document.body);
 };