+++ platforms/uni-app/src/index.ts
@@ -1,0 +1,4 @@
+export * from '@vtj/renderer';
+export * from '@vtj/uni';
+export * from '@vtj/core';
+export * from './modules';