+++ create-vtj/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.9.11](https://gitee.com/newgateway/vtj/compare/create-vtj@0.9.10...create-vtj@0.9.11) (2025-02-13)
+
+**Note:** Version bump only for package create-vtj
+
+
+
+
+
 ## [0.9.10](https://gitee.com/newgateway/vtj/compare/create-vtj@0.9.9...create-vtj@0.9.10) (2025-02-13)